pyproject.toml
[tool.poetry]
name = "umlizer"
version = "0.1.1"  # semantic-release
description = "Tool for creating UML from project files"
authors = ["Ivan Ogasawara <ivan.ogasawara@gmail.com>"]
license = "BSD-3-Clause"
readme = "README.md"
exclude = [
  ".git/*",
  ".env*",
]
packages = [
  {include = "umlizer", from="src"},
]
include = ["src/umlizer/py.typed"]

[tool.poetry.scripts]
"umlizer" = "umlizer.__main__:app"

[tool.poetry.dependencies]
python = ">=3.8.1,<4"
graphviz = ">=0.20.1"
atpublic = ">=4.0"
typing-extensions = { version = ">=4", python = "<3.9" }
typer = ">=0.9.0"
pyyaml = ">=5.4"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.3.2"
pytest-cov = ">=4.1.0"
coverage = ">=7.2.7"
pre-commit = ">=3.3.2"
ruff = ">=0.1.5"
mypy = ">=1.6.0"
bandit = ">=1.7.5"
vulture = ">=2.9.1"
mccabe = ">=0.6.1"
ipykernel = ">=6.0.0"
ipython = ">=7.0.0"
mkdocs = ">=1.4.3"
mkdocs-exclude = ">=1.0.2"
mkdocs-jupyter = ">=0.24.1"
mkdocs-literate-nav = ">=0.6.0"
mkdocs-macros-plugin = ">=0.7.0,<1"
mkdocs-material = ">=9.1.15"
mkdocstrings = ">=0.21.2"
mkdocstrings-python = ">=1.1.2"
jupyterlab = ">=4.0.5"
makim = "1.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
testpaths = [
    "tests",
]

[tool.bandit]
exclude_dirs = ["tests"]
targets = "./"
skips = ["B602"]

[tool.vulture]
exclude = ["tests"]
ignore_decorators = ["@abc.abstractmethod"]
ignore_names = []
make_whitelist = true
min_confidence = 80
paths = ["./"]
sort_by_size = true
verbose = false

[tool.ruff]
line-length = 79
force-exclude = true
src = ["./"]
ignore = ["PLR0913", "RUF008"]
exclude = [
  'docs',
]
select = [
  "E",   # pycodestyle
  "F",   # pyflakes
  "D",   # pydocstyle
  "YTT", # flake8-2020
  "PL",  # PL
  "RUF", # Ruff-specific rules
  "I001", # isort
]
fix = true

[tool.ruff.pydocstyle]
convention = "numpy"

[tool.ruff.isort]
# Use a single line between direct and from import
lines-between-types = 1

[tool.ruff.format]
quote-style = "single"

[tool.mypy]
python_version = "3.8"
check_untyped_defs = true
strict = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true


[[tool.mypy.overrides]]
module = [
  "yaml",
]
ignore_missing_imports = true
